clinic/requirements.txt
awesome-slugify==1.6.5
Babel==2.5.3
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
Django==1.11
django-embed-video==1.1.2
django-htmlmin==0.10.0
django-localized-fields==4.5
django-modeltranslation==0.12.2
django-phonenumber-field==2.0.0
django-postgres-extra==1.20
django-widget-tweaks==1.4.2
html5lib==1.0.1
idna==2.6
lxml==4.2.1
phonenumberslite==8.9.4
Pillow==5.1.0
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.4
regex==2018.2.21
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
webencodings==0.5.1
